refactor: rename InputOrGroup to AutoField
