+++ apps/docs/pages/docs/api-reference/components/field-label.mdx
@@ -3,7 +3,7 @@ title: <FieldLabel>
 ---
 
 import { ConfigPreview } from "../../../../components/Preview";
-import { FieldLabel } from "@/core/components/InputOrGroup";
+import { FieldLabel } from "@/core/components/AutoField";
 import { Globe } from "lucide-react";
 
 # \<FieldLabel\>
