+++ apps/docs/pages/docs/extending-puck/custom-fields.mdx
@@ -1,5 +1,5 @@
 import { ConfigPreview } from "@/docs/components/Preview";
-import { FieldLabel } from "@/core/components/InputOrGroup";
+import { FieldLabel } from "@/core/components/AutoField";
 
 # Custom Fields
